+++ mirrulations-client/.coveragerc
@@ -1,0 +1,4 @@
+[report]
+exclude_lines =
+    pragma: no cover
+    if __name__ == .__main__.:
